chore: removes mounting support to be implemented in separate PR
