+++ src/OrasProject.Oras/CopyOptions.cs
@@ -1,0 +1,39 @@
+// Copyright The ORAS Authors.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using OrasProject.Oras.Oci;
+
+namespace OrasProject.Oras;
+
+public struct CopyOptions
+{
+    public event Action<Descriptor> OnPreCopy;
+    public event Action<Descriptor> OnPostCopy;
+    public event Action<Descriptor> OnCopySkipped;
+
+    internal void PreCopy(Descriptor descriptor)
+    {
+        OnPreCopy?.Invoke(descriptor);
+    }
+
+    internal void PostCopy(Descriptor descriptor)
+    {
+        OnPostCopy?.Invoke(descriptor);
+    }
+
+    internal void CopySkipped(Descriptor descriptor)
+    {
+        OnCopySkipped?.Invoke(descriptor);
+    }
+}
